straight move of workload controller
