+++ pkg/operator/workloadcontroller/workload_controller_openshiftapiserver_v311_00.go
@@ -1,4 +1,4 @@
-package operator
+package workloadcontroller
 
 import (
 	"fmt"
